Refactoring flat_install
